Save and display quoted tweets.
If a Tweet quotes another, and has that data in its JSON, we save that.
It can be found at `Tweet.get_quoted_tweet()`.
We also display it in the templates.

Re-jigged how the `TweetMixin` and `UserMixin` work slightly, so that it's easier
to save the quoted tweets. Now, `TweetMixin` inherits `UserMixin`, so it can save
a Tweet's user, which seems sensible, and simplifies some other things.

Also: Stopped using `profile_image_url` - now only use `profile_image_url_https`.
But there is a `@property` on Users called `profile_image_url`, which returns
`profile_image_url_https`.

Also, we store `favourite_count` spelt like that, as that's the way it's spelt
in the API. But a `@property` will answer to `favorite_count`.

Fixes #39