Last.fm views, templates and loads of other terrible changes.
Yeah, became a bit of a monster.

So we have some basic views and templates for Last.fm.

But I also ended up having to rework lots of the fetching, at least
twice, and re-setting the migrations, at least twice, and changing
the URL scheme, and fixing some project-wide tests, and storing
the app name and version in `ditto/__init__.py` so it can be
used from within the code, and what a mess.