Add lowercase slugs to Artists, Tracks and Albums.
We need to match imported things case insensitively, so that if
a slug simply changes its capitalization at Last.fm in between
fetches, we don't think it's a new Artist or whatever.

But SQLite can't do case-insensitive matching of Unicode strings.

So we've ended up with each thing having:

    * `original_slug` - As used on Last.fm, so we can link to its
      original page there.
    * `slug` - A lowercase version which we use for (a) matching
      entities when we fetch them and (b) for URLs on Ditto.

Seeing as we have lowercase slugs, seems sensible to use them.