Fix Flickr import error
When importing a tag that's like slug='phil-gyford', name='Phil Gyford'
if there was already a tag that's slug='philgyford', name='Phil Gyford'
we were creating a new one, as the slug's different. But that would
create an IntegrityError because there can't be duplicate slugs.

So now we catch that and, if it happens, use the existing tag.