Fix fetching of extended tweets and handling 280 character tweets
For some reason we weren't fetching extended tweet texts in
`FetchTweetsRecent._call_api()`. We now are. I hope that's OK.

Also, the `title` and `summary` of saved `Tweet` objects (which are
largely pointless) weren't being truncated. This "worked" quietly
on SQLite, because it doesn't enforce their 255 character limits,
but broke on Postgres as nature intended.

So now we manually truncate both strings before trying to save the
Tweet.

ALSO, we now force devproject and the tox tests to use a specific
commit of Twython because that has a fix for a bug in it which isn't
yet in the version of Twython on pypi (3.6.0).