Fix media file paths in tests for Django 3.1
I hadn't set `MEDIA_URL` in the test settings file. If this isn't
explicitly set then (in the absence of a server-provided `SCRIPT_NAME`)
then `settings.MEDIA_URL` is "" (empty string) in Django 3.0 (and, I
guess, 2.2) but "/" (forward slash) in Django 3.1.

So, set `MEDIA_URL` to `/media/` in the test settings and update
the expected paths in the tests, and they now pass in Django 2.2, 3.0
and 3.1.