Fix lots and lots of tests
Mostly as a result of me doing stuff that Ruff said, but doing it
wrongly.

A few instances of datetimes, now with timezones, that were out, but
which I *think* are now correct.

And a few instances where a templatetag was being called with an
arg instead of a kwarg