Set default of all post_time fields to now
The `post_time` field on all "item" models – `Photo`, `Scrobble`,
`Bookkark` and `Tweet` records when the thing was posted on the
service. Its value is set when the data is imported.

But if someone wanted to, e.g. start using the Pinboard app to save
Bookmarks locally, without importing from Pinboard, then it's a pain
to not have "now" set, given it's a required field.

So this field is set to "now" by default, but can be changed.